requirements.txt
openai-whisper
pyaudio
mistralai